+++ nomos-core/src/main/scala/dev/cjfravel/nomos/generation/TemplateSerializer.scala
@@ -25,6 +25,7 @@ object TemplateSerializer {
       fromJsonStyle = "${escapeString(template.fromJsonStyle)}",
       dateType = "${escapeString(template.dateType)}",
       dateTimeType = "${escapeString(template.dateTimeType)}",
+      mapType = "${escapeString(template.mapType)}",
       visibility = $visibility
     )"""
   }